requirements.txt
gradio==3.30.0
gradio_client==0.2.4
huggingface-hub==0.14.1
imageio==2.25.1
imgcat==0.5.0
ipykernel==6.16.0
ipython==8.5.0
ipywidgets==8.0.2
kiwisolver==1.4.2
kornia==0.6.9
legacy==0.1.6
numpy==1.21.6
omegaconf==2.2.3
opencv-python==4.5.5.62
opencv-python-headless==4.7.0.68
packaging==21.3
pandas==1.4.2
parso==0.8.3
Pillow==9.4.0
protobuf==3.20.2
Pygments==2.13.0
PyMatting==1.1.8
pyparsing==3.0.9
pyrsistent==0.19.3
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.10.0
seaborn==0.12.2
sklearn==0.0
sniffio==1.3.0
soupsieve==2.4
timm==0.6.12
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
transformers==4.28.1
onnx==1.14.1
onnxruntime==1.16.0